# persist: handle trace compaction outside of main Indexed loop.