--- a/src/persist/src/indexed/metrics.rs
+++ b/src/persist/src/indexed/metrics.rs
@@ -84,6 +84,7 @@ pub struct Metrics {
     pub(crate) compaction_count: ThirdPartyMetric<UIntCounter>,
     pub(crate) compaction_seconds: ThirdPartyMetric<Counter>,
     pub(crate) compaction_write_bytes: ThirdPartyMetric<UIntCounter>,
+    pub(crate) trace_compaction_error_response_count: ThirdPartyMetric<UIntCounter>,
 
     // TODO: Tag cmd_process_count with cmd type and remove this?
     pub(crate) cmd_write_count: ThirdPartyMetric<UIntCounter>,
@@ -167,6 +168,10 @@ impl Metrics {
             compaction_write_bytes: registry.register_third_party_visible(metric!(
                 name: "mz_persist_compaction_bytes",
                 help: "bytes written compacting unsealed and trace",
+            )),
+            trace_compaction_error_response_count: registry.register_third_party_visible(metric!(
+                name: "mz_persist_trace_compaction_error_response_count",
+                help: "count of trace compaction requests where an error was returned",
             )),
             cmd_write_count: registry.register_third_party_visible(metric!(
                 name: "mz_persist_cmd_write_count",