<!--

Describe the contents of the PR briefly but completely.

If you write detailed commit messages, it is acceptable to copy/paste them
here, or write "see commit messages for details." If there is only one commit
in the PR, GitHub will have already added its commit message above.

-->

### Motivation

This PR adds the long desired feature of handling trace compaction work outside of the main Indexed loop so that it doesn't affect write latency/throughput.

This PR broadly follows a lot of the structure from #8472.

At a high level, instead of Indexed performing trace compactions periodically in `step`, Indexed hands out a list of `PFuture`s representing the compaction work. `RuntimeImpl` waits for those to resolve asynchronously, and when they do,
they return their output by sending a new command back to `RuntimeImpl`'s queue. When `RuntimeImpl` receives a command indicating that some trace compaction was successfully completed, it applies the trace compaction as a batched command, and when the Meta for the sequence of batched commands is successfully committed, it physically deletes the trace batches that are no longer needed. If meta fails to commit, no worries, the state reverts to the previous meta and we start over again. Indexed is careful to only keep one trace compaction request in flight per collection at a time.

Individual commits have more detail and I really tried to keep commits small here for ease of review :)

<!--

Which of the following best describes the motivation behind this PR?

  * This PR fixes a recognized bug: [Link to issue.]

  * This PR adds a known-desirable feature: [Link to issue.]

  * This PR fixes a previously unreported bug.

    [Describe the bug in detail, as if you were filing a bug report.]

  * This PR adds a feature that has not yet been specified.

    [Write a brief specification for the feature, including justification
     for its inclusion in Materialize, as if you were writing the original
     feature specification.]

   * This PR refactors existing code.

    [Describe what was wrong with the existing code, if it is not obvious.]

-->

### Tips for reviewer

<!--

Leave some tips for your reviewer, like:

    * The diff is much smaller if viewed with whitespace hidden.
    * [Some function/module/file] deserves extra attention.
    * [Some function/module/file] is pure code movement and only needs a skim.

Delete this section if no tips.

-->

### Checklist

- [x] This PR has adequate test coverage / QA involvement has been duly considered.
- [ ] This PR adds a release note for any [user-facing behavior changes](https://github.com/MaterializeInc/materialize/blob/main/doc/user/content/release-notes.md#what-changes-require-a-release-note).

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/materializeinc/materialize/9602)
<!-- Reviewable:end -->