@ruchirK 2021-12-15
RFAL! still todo here is:

perf testing to make sure this is a perf improvement / maybe also write a ubenchmark
unit testing -- haven't thought about exactly whats the best way to test this or how it interacts with existing unit tests yet.

but I think this works and is fairly clean.