@ruchirK 2021-12-15
yeah! I plan to spend today trying to see if I can make that work! I definitely think thats a more ideal end state.