@danhhz 2021-12-16
okay cool, I was imagining that this last bit would be straightforward. in this pr is my preference. i'll take a look at this today