@ruchirK 2021-12-22
I think I have a few more tests to write:

* make sure compaction proceeds as expected if a meta write fails
* make sure compaction proceeds as expected if it receives a error/compactionres response when there are no pending requests in flight.
* make sure everything is fine if it receives a response for an id that is neither deleted nor created

and i think I might have one or two comments that I missed but otherwise this is getting close!