@ruchirK 2021-12-27
This is RFAL!

This is RFAL!

I think I've addressed all comments except for the "add tests for the multi-collection case" which I'll do in this PR after the next round of comments. I kept this one to 3 commits - each of the first two introduce ~7 lines of noisiness that is taken care of in the subsequent commit. The last commit is fairly large, but I couldn't think of any way to break it up further that wouldn't also be extremely noisy. I'm open to any suggestions for that. Also, a lot of it is test code.


I think the only controversial bits now are:

I decided to keep the check that the response comes from the expected request, and added `soft_assert_or_log!`s around the cases when we either get a request when there isn't any in flight, or when we receive a response from an unexpected request. I left comments in the code and in a thread in this PR explaining why I think thats a better option for now. Unfortunately, the soft assertions makes testing this more convoluted as e.g. invalid responses that don't come from the expected request now panic -- I worked around that with `catch_unwind` - let me know if there's a better way or if the soft_asserts don't seem worth it.