@ruchirK 2021-12-28
Morning! I broke the third commit into two commits -- the first one simply teaches persist to send MaintenanceReqs and handles MaintenanceResponses as a batched cmd, and then the second teaches persist to only have one request in flight at a time. Hopefully this makes for easier review.

I also figured out another reason why its important to check that the response matches the request we have in flight, and only use it if so -- otherwise, we risk breaking the invariant that trace batch compaction levels are nonincreasing.

Copying from the new commit message:

```
    For example, say a stream has trace batches

    [a, b, c, d, e, f] and currently has a compaction request out for batches

    (c, d). This implies that (a, b) and (b, c) were not compactable because if they
    had been, we would have sent a trace compaction request for those instead (the
    selection algorithm simply scans pairs left to right looking for two trace batches
    not in advance of the compaction frontier that are at the same compaction level).

    At this point, if we receive a trace compaction response for (e, f) -> g,
    g can be at a new compaction level and if we use that result, turning the list
    of trace batches into

    [a, b, c, d, g], we can potentially break the invariant that compaction levels are
    nonincreasing (if say, a, b, c, d were all at the same compaction level).
```
I think it might be time to change that invariant but I would much rather do that in a followup commit.