@ruchirK 2021-12-28
No worries -- can you take a quick look at the very first commit here (splitting up trace_compact into two functions and pulling Maintainer out of arrangement) and if that's seems fine to rubber stamp I'd love to pull that out and get that merged (really anything to pull down the size of this change). No worries if not