@ruchirK 2022-01-19
@danhhz the latest commit I just pushed removes all of the `soft_assert_or_log`s and replaces them with the approach we had discussed last week. Lmk if that all looks good, and then I'll clean this up and get it ready for finally merging!