@ruchirK 2022-01-26
Ok whew! This pr is cleaned up and ready for what I hope is a final look before merging. I had to rewrite a lot of the older commits to get everything to compile, but the most recent set of responses to everyone's comments are in the commit titled
`fixup! respond to code review comments v2`

If there are no objections I'd most prefer to merge this all as one large commit as I've tried to find good boundaries for commits for a while now, and everything I come up with seems either trivial (e.g. 2 tiny commits, one huge commit with 95% of the diff, or commits that introduce a lot of churn amongst themselves)

For the performance analysis: I'm going to get the microbenchmark + feature benchmark + open loop benchmark for ~1 - 2 minutes of running at 10k with just these changes, but then I'm going to compare these changes against main, with this diff applied to both:

```
diff --git a/src/persist/src/operators/stream.rs b/src/persist/src/operators/stream.rs
index 8593d17a8..ee8282fc9 100644
--- a/src/persist/src/operators/stream.rs
+++ b/src/persist/src/operators/stream.rs
@@ -314,6 +314,11 @@ where

                             // Explicitly downgrade the capability to the new time.
                             cap_set.downgrade(Some(pending_future.time));
+                            if pending_future.time > 1000 {
+                                let _ = write.allow_compaction(Antichain::from_elem(
+                                    pending_future.time - 1000,
+                                ));
+                            }
                         }
                         std::task::Poll::Ready(Err(e)) => {
                             tracing::trace!(
altaria-2:persist Test$
```

This diff hardcodes a compaction interval of 1s into the seal operator and can be used as a relatively simple testbed of "how does this change affect performance of kafka sources with compaction enabled"