@ruchirK 2022-01-27
Ok now time for feature benchmark results. As promised, I ran the benchmarks across 3 different configurations, comparing all 3 against `v0.18.0` as a stable baseline (it seemed difficult to compare all three against each other directly). Each test ran with

```
bin/mzcompose --find feature-benchmark run default --other-tag v0.18.0 --this-options "--persistent-user-tables --persistent-kafka-upsert-source" --other-options "--persistent-user-tables --persistent-kafka-upsert-source"
```
on 3 separate r5a.4xlarge machines, each of which was running nothing else significant. These tests were all done on a separate detached `screen` session so they wouldn't be affected if e.g. my laptop died. 

The configurations were:

1. The changes in this PR, without any compaction added. The hypothesis here was that this should mostly be a no-op relative to v0.18.0 because the persistent kafka sources don't enable any compaction and so this would perhaps show up as a speedup for tests that insert lots of data into tables.
2. `main` (`d10ac1e556d2edb7221cad0146ca041c364273a0`) + the diff I described earlier to hardcode logical compaction to 1second relative to the most recent seal timestamp. The hypothesis was that this should be a strict slowdown relative to v0.18.0 because now main will be doing compactions on the Indexed loop. Note btw that `KafkaRecovery` is all unique (k,v) pairs so we won't get the benefit of data reduction when we enable compaction.
3. The changes in this pr + the diff to hardcode compactions. The hypothesis for this change was that it should be somewhere between the two configurations described above. It makes sense for it to perhaps be slower than just this PR without compaction hardcoded in (because compactions do introduce extra CPU and IO work to be done _somewhere_), but the impact of that additional work should be much less observable to someone running queries against Materialize. 

Ok, so onto the results.

1. As expected, the changes in this PR were all within +/-5% of the results in v0.18.0. `Insert` and `Update` were `2.1` and `0.8` percent faster respectively. The raw data is [pr-nocompaction.txt](https://github.com/MaterializeInc/materialize/files/7947120/pr-nocompaction.txt).

2. Somewhat surprisingly, running main with hardcoded compaction did not result in as much of a slowdown as I had remembered. Zooming in on the relevant benchmarks I care most about (main with compaction is the `this` column, `v0.18.0` is the `other` column)

```
KafkaUpsert               |       1.874 |       1.942 |      no       | 3.5 pct   faster
KafkaUpsertUnique         |       2.040 |       1.991 |      no       | 2.5 pct   slower
KafkaRecovery             |      20.659 |      19.443 |      no       | 6.3 pct   slower
ExactlyOnce               |       6.012 |       5.522 |      no       | 8.9 pct   slower
```

There's slight slowdowns here but I was expecting something on the order of 30 - 40%, especially for `KafkaRecovery` and `ExactlyOnce`, which both use upsert kafka sources and run for a relatively long time. The raw data file is 
[main-compaction.txt](https://github.com/MaterializeInc/materialize/files/7947232/main-compaction.txt)

I reran just `KafkaRecovery` with

```
bin/mzcompose --find feature-benchmark run default --other-tag v0.18.0 --this-options "--persistent-user-tables --persistent-kafka-upsert-source" --other-options "--persistent-user-tables --persistent-kafka-upsert-source" --root-scenario KafkaRecovery
```

and got drastically different results

```
NAME                      |    THIS     |    OTHER    |  Regression?  | 'THIS' is:
----------------------------------------------------------------------------------------------------
KafkaRecovery             |      29.088 |      19.175 |    !!YES!!    | 51.7 pct   slower
```
The raw data file is 
[main-compaction-kafka-recovery.txt](https://github.com/MaterializeInc/materialize/files/7947249/main-compaction-kafka-recovery.txt)

Just to triple check myself, I re-ran _again_ the exact same code, the exact same test, on the exact same machine, and got a totally different result again

```
NAME                      |    THIS     |    OTHER    |  Regression?  | 'THIS' is:
----------------------------------------------------------------------------------------------------
KafkaRecovery             |      20.411 |      19.266 |      no       | 5.9 pct   slower
```

The raw data file for this run is: 
[main-compaction-kafka-recovery-2.txt](https://github.com/MaterializeInc/materialize/files/7947279/main-compaction-kafka-recovery-2.txt)


Analysis: if you look at the two individual `KafkaRecovery` runs, the 51.7% slower result is _not_ a one-off fluke. There's multiple runs that take 35 seconds or more, and switching to v0.18.0, all runs take around 20 seconds again. But in the second run, both main + hardcoded compaction and and v0.18.0 take 20 +/- 2 seconds! There's nothing obvious I can point to to explain the observed behavior. My takeaway here is that this benchmark is not really going to give a high confidence yes or no to the changes in this branch, and we'll have to use the open loop test to try to suss that out. It's also possible that there's some user error at play here.

3. Running the changes in this PR + compaction hardcoded produced a slowdown consistent with main + compaction hardcoded

```
KafkaUpsert               |       1.938 |       1.936 |      no       | 0.1 pct   slower
KafkaUpsertUnique         |       1.977 |       1.964 |      no       | 0.7 pct   slower
KafkaRecovery             |      21.439 |      19.294 |    !!YES!!    | 11.1 pct   slower
ExactlyOnce               |       5.686 |       4.557 |    !!YES!!    | 24.8 pct   slower
```

Raw data file is: 
[pr-compaction.txt](https://github.com/MaterializeInc/materialize/files/7947308/pr-compaction.txt)

Rerunning just `KafkaRecovery`:

```
NAME                      |    THIS     |    OTHER    |  Regression?  | 'THIS' is:
----------------------------------------------------------------------------------------------------
KafkaRecovery             |      20.919 |      19.130 |      no       | 9.3 pct   slower
```
Raw data file is: 
[pr-compaction-kafka-recovery.txt](https://github.com/MaterializeInc/materialize/files/7947321/pr-compaction-kafka-recovery.txt)

Analysis: these results seem entirely consistent with the variance already observed in repeating this benchmark multiple times, and we cannot conclude whether this code is better or worse based on these results.

I'm switching gears to the open loop benchmark. Laying out my hypothesis before I do the experiment in the interest of good science: I fully expect that one to show a clear difference in how much throughput this PR+compaction hardcoded handle vs main+compaction hardcoded can handle.